use Twitter's async media upload
switch to using the newer async media upload process, which require an INIT, APPPEND, FINALIZE, and optional STATUS query. it's a bunch more work but apparently is the model Twitter plans to force eventually.

the bulk of the work is still done by signedPostImage, which is unchanged. the upload and updateWithMedia calls have to take the mime type and category now.